+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Fixes #58: Guess command not working.
 - Fixes #56: Unable to use glide get on golang.org/x/[name]/[subpackage]
+- Fixes #61: The wrong version of a dependency can be pinned when packages are
+  vendored (no VCS repo associated with them).
 
 # Release 0.5.0 (2015-08-19)